V2/requirements.txt
pdfplumber
icalendar
pandas
pytest
pytz
google-api-python-client
google-auth-httplib2
google-auth-oauthlib